Fix warnings from vs2010
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@1516630 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/test-integration/activemq/test/CmsSendWithAsyncCallbackTest.cpp
@@ -91,7 +91,7 @@ namespace {
                 dynamic_cast<ActiveMQProducer*>(session->createProducer(destination.get())));
         producer->setDeliveryMode(DeliveryMode::PERSISTENT);
 
-        long start = System::currentTimeMillis();
+        long long start = System::currentTimeMillis();
 
         for (int i = 0; i < count; i++) {
             std::auto_ptr<cms::TextMessage> message(session->createTextMessage("Hello"));
@@ -110,7 +110,7 @@ namespace {
                 dynamic_cast<ActiveMQProducer*>(session->createProducer(destination.get())));
         producer->setDeliveryMode(DeliveryMode::PERSISTENT);
 
-        long start = System::currentTimeMillis();
+        long long start = System::currentTimeMillis();
 
         for (int i = 0; i < count; i++) {
             std::auto_ptr<cms::TextMessage> message(session->createTextMessage("Hello"));
